An unresolved symlink in Coda shows as an empty entry.linkname
Ignore the symlink until resolved, we'll back up from another replica.

+++ deltaic/sources/coda.py
@@ -162,7 +162,7 @@ def update_dir_from_tar(tar, root_dir):
             if update_file(path, TarMemberFile(tar, entry)):
                 print 'f', path
         elif entry.issym():
-            if st is None or os.readlink(path) != entry.linkname:
+            if st is None or entry.linkname and os.readlink(path) != entry.linkname:
                 print 's', path
                 if st is not None:
                     os.unlink(path)
